# sources/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/item-categories": {
      "post": {
        "operationId": "ItemCategoriesController_create",
        "summary": "Создание записи",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateItemCategoryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Запись создана успешно",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Не верный формат данных в JSON",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Access токен просрочен \n\nAccess токен не передан \n\nУ пользователя нет роли администратора \n\n",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "В БД уже есть запись с таким полем",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Что-то пошло не так на сервере.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "api_v1_item-categories",
          "ADMIN"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      },
      "get": {
        "operationId": "ItemCategoriesController_findAll",
        "summary": "Получить список записей",
        "parameters": [
          {
            "name": "dp_itemBrandId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "brand",
            "required": false,
            "in": "query",
            "example": "de-pa-electric",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Получили список записей",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ItemCategoryWithIdDto"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Что-то пошло не так на сервере.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "api_v1_item-categories",
          "any"
        ]
      }
    },
    "/api/v1/item-categories/bulk": {
      "post": {
        "operationId": "ItemCategoriesController_createBulk",
        "summary": "Создание несколько записей за раз",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBulkItemCategoryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Записи добавлены в БД",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Не верный формат данных в JSON",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Access токен просрочен \n\nAccess токен не передан \n\nУ пользователя нет роли администратора \n\n",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "В БД уже есть запись с таким полем",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Что-то пошло не так на сервере.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "api_v1_item-categories",
          "ADMIN"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      },
      "put": {
        "operationId": "ItemCategoriesController_updateBulk",
        "summary": "Обновление записей транзакцией",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBulkItemCategoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Обновили записи. Транзакция выполнена.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Не верный формат данных в JSON",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Access токен просрочен \n\nAccess токен не передан \n\nУ пользователя нет роли администратора \n\n",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Запись не найдена",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "В БД уже есть запись с таким полем",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Что-то пошло не так на сервере.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "api_v1_item-categories",
          "ADMIN"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/api/v1/item-categories/{id}": {
      "get": {
        "operationId": "ItemCategoriesController_findOne",
        "summary": "Получить запись по id",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Получили запись по id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemCategoryWithIdDto"
                }
              }
            }
          },
          "404": {
            "description": "Запись не найдена",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Что-то пошло не так на сервере.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "api_v1_item-categories",
          "any"
        ]
      },
      "patch": {
        "operationId": "ItemCategoriesController_update",
        "summary": "Обновление записи по id",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateItemCategoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Обновили запись по id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Не верный формат данных в JSON",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Access токен просрочен \n\nAccess токен не передан \n\nУ пользователя нет роли администратора \n\n",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Запись не найдена",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "В БД уже есть запись с таким полем",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Что-то пошло не так на сервере.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "api_v1_item-categories",
          "ADMIN"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      },
      "delete": {
        "operationId": "ItemCategoriesController_remove",
        "summary": "Удаление записи по id",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Удалили запись по id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Access токен просрочен \n\nAccess токен не передан \n\nУ пользователя нет роли администратора \n\n",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Запись не найдена",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Что-то пошло не так на сервере.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "api_v1_item-categories",
          "ADMIN"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
...
    "/api/v1/items/bulk": {
      "post": {
        "operationId": "ItemsController_createBulk",
        "summary": "Создание несколько записей за раз транзакцией",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBulkItemDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Записи добавлены в БД. Транзакция выполнена.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Не верный формат данных в JSON",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Access токен просрочен \n\nAccess токен не передан \n\nУ пользователя нет роли администратора \n\n",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "В БД уже есть запись с таким полем",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Транзакция не выполнена.Что-то пошло не так на сервере.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "api_v1_items",
          "ADMIN"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      },
      "put": {
        "operationId": "ItemsController_updateBulk",
        "summary": "Обновление записи по uuid транзакцией",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBulkItemDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Обновили запись по uuid. Транзакция выполнена.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Не верный формат данных в JSON",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Access токен просрочен \n\nAccess токен не передан \n\nУ пользователя нет роли администратора \n\n",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Запись не найдена",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "В БД уже есть запись с таким полем",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Транзакция не выполнена.Что-то пошло не так на сервере.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "api_v1_items",
          "ADMIN"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
...
    "/api/v1/items/{id}": {
      "get": {
        "operationId": "ItemsController_findOne",
        "summary": "Получить запись по id",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Получили запись по id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemEntity"
                }
              }
            }
          },
          "404": {
            "description": "Запись не найдена",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Что-то пошло не так на сервере.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "api_v1_items",
          "any"
        ]
      },
      "patch": {
        "operationId": "ItemsController_update",
        "summary": "Обновление записи по uuid транзакцией",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateItemDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Обновили запись по uuid. Транзакция выполнена.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Не верный формат данных в JSON",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Access токен просрочен \n\nAccess токен не передан \n\nУ пользователя нет роли администратора \n\n",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Запись не найдена",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "В БД уже есть запись с таким полем",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Транзакция не выполнена.Что-то пошло не так на сервере.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "api_v1_items",
          "ADMIN"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      },
      "delete": {
        "operationId": "ItemsController_remove",
        "summary": "Удаление записи по id",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Удалили запись по id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Access токен просрочен \n\nAccess токен не передан \n\nУ пользователя нет роли администратора \n\n",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Запись не найдена",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Что-то пошло не так на сервере.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "api_v1_items",
          "ADMIN"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "REST API документация",
...
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "access token - токен доступа",
        "in": "header"
      },
      "refresh-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "refresh token - токен обновления",
        "in": "header"
      }
    },
    "schemas": {
...
      "HttpResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 0
          },
          "message": {
            "type": "string",
            "example": "сообщение"
          }
        },
        "required": [
          "statusCode",
          "message"
        ]
      },
...
      "CreateItemCategoryDto": {
        "type": "object",
        "properties": {
          "dp_name": {
            "type": "string",
            "example": "Аксессуары"
          },
          "dp_sortingIndex": {
            "type": "number",
            "example": 10000
          },
          "dp_urlSegment": {
            "type": "string",
            "example": "dp-accessories"
          },
          "dp_photoUrl": {
            "type": "string",
            "example": "https://ooodepa.github.io/ooodepa_imgs/products/de-pa-electric/accessories/11202.png"
          },
          "dp_seoKeywords": {
            "type": "string",
            "example": ""
          },
          "dp_seoDescription": {
            "type": "string",
            "example": ""
          },
          "dp_isHidden": {
            "type": "boolean",
            "example": false
          },
          "dp_itemBrandId": {
            "type": "number",
            "example": 0
          }
        },
        "required": [
          "dp_name",
          "dp_sortingIndex",
          "dp_urlSegment",
          "dp_photoUrl",
          "dp_seoKeywords",
          "dp_seoDescription",
          "dp_isHidden"
        ]
      },
...
      "CreateBulkItemCategoryDto": {
        "type": "object",
        "properties": {
          "bulk": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ItemCategoryNoIdDto"
            }
          }
        },
        "required": [
          "bulk"
        ]
      },
      "ItemCategoryWithIdDto": {
        "type": "object",
        "properties": {
          "dp_name": {
            "type": "string",
            "example": "Аксессуары"
          },
          "dp_sortingIndex": {
            "type": "number",
            "example": 10000
          },
          "dp_urlSegment": {
            "type": "string",
            "example": "dp-accessories"
          },
          "dp_photoUrl": {
            "type": "string",
            "example": "https://ooodepa.github.io/ooodepa_imgs/products/de-pa-electric/accessories/11202.png"
          },
          "dp_seoKeywords": {
            "type": "string",
            "example": ""
          },
          "dp_seoDescription": {
            "type": "string",
            "example": ""
          },
          "dp_isHidden": {
            "type": "boolean",
            "example": false
          },
          "dp_itemBrandId": {
            "type": "number",
            "example": 0
          },
          "dp_id": {
            "type": "number",
            "example": 0
          }
        },
        "required": [
          "dp_name",
          "dp_sortingIndex",
          "dp_urlSegment",
          "dp_photoUrl",
          "dp_seoKeywords",
          "dp_seoDescription",
          "dp_isHidden",
          "dp_id"
        ]
      },
      "UpdateItemCategoryDto": {
        "type": "object",
        "properties": {
          "dp_name": {
            "type": "string",
            "example": "Аксессуары"
          },
          "dp_sortingIndex": {
            "type": "number",
            "example": 10000
          },
          "dp_urlSegment": {
            "type": "string",
            "example": "dp-accessories"
          },
          "dp_photoUrl": {
            "type": "string",
            "example": "https://ooodepa.github.io/ooodepa_imgs/products/de-pa-electric/accessories/11202.png"
          },
          "dp_seoKeywords": {
            "type": "string",
            "example": ""
          },
          "dp_seoDescription": {
            "type": "string",
            "example": ""
          },
          "dp_isHidden": {
            "type": "boolean",
            "example": false
          },
          "dp_itemBrandId": {
            "type": "number",
            "example": 0
          }
        },
        "required": [
          "dp_name",
          "dp_sortingIndex",
          "dp_urlSegment",
          "dp_photoUrl",
          "dp_seoKeywords",
          "dp_seoDescription",
          "dp_isHidden"
        ]
      },
      "UpdateBulkItemCategoryDto": {
        "type": "object",
        "properties": {
          "bulk": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ItemCategoryWithIdDto"
            }
          }
        },
        "required": [
          "bulk"
        ]
      },
...
      "CreateBulkItemDto": {
        "type": "object",
        "properties": {
          "bulk": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ItemNoIdDto"
            }
          }
        },
        "required": [
          "bulk"
        ]
      },
...
      "UpdateBulkItemDto": {
        "type": "object",
        "properties": {
          "bulk": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ItemWithIdDto"
            }
          }
        },
        "required": [
          "bulk"
        ]
      },
...
    }
  }
}